+++ demos/polar/colormap.py
@@ -43,7 +43,7 @@ from mazes.Algorithms.hunt_kill import HuntKill
 from mazes.tools.distance_map import DistanceColoring
 
     # graphics
-from mazes.Graphics.polar1 import Phocidae
+from mazes.Graphics.polar1 import Pholcidae
 
 def init_maze(rings:int, pole:int=6, split:int=1) -> Maze:
     """create an empty theta maze"""
@@ -95,7 +95,7 @@ def main(grid_args:tuple, graphics_args:tuple, alg_args,
     distances = DistanceColoring(maze,
                                  to_rgb(hot), to_rgb(cold), to_rgb(zero),
                                  source=start_cell)
-    spider = Phocidae(maze)
+    spider = Pholcidae(maze)
     spider.setup(fillcolors=distances.gradients)
     spider.title(title)
     spider.draw_maze()