+++ demos/vgt_bfs.py
@@ -25,7 +25,7 @@ LICENSE
 """
 import mazes
 from mazes.VGT.bfs import bfs, init_maze
-from mazes.Graphics.oblong1 import Phocidae
+from mazes.Graphics.oblong1 import Pholcidae
 
 def make_maze(maze:"Maze", start_cell:"Cell", shuffle:bool=True) -> "Status":
     """carve the maze"""
@@ -43,7 +43,7 @@ def main(dim, start, shuffle, color, output, console, gui):
     if console or (dim[0] <= 10 and dim[1] <= 15):
         print(maze)
 
-    spider = Phocidae(maze)
+    spider = Pholcidae(maze)
     spider.setup(color=color)
     title = "Breadth-first search "
     if shuffle: