+++ mazes/Graphics/oblong1.py
@@ -9,7 +9,7 @@ DESCRIPTION
     and the spider graphics driver.  It does not support one-way connections
     or insets. One class is defined:
 
-        Phocidae -- named after the family of daddy long-leg spiders.
+        Pholcidae -- named after the family of daddy long-leg spiders.
 
 LICENSE
 
@@ -32,12 +32,15 @@ MODIFICATIONS
         Added cell fill with color.
     28 December 2024 - EC
         Correct a typo in the docstring.
+    25 October 2025 - EC
+        Correct the spelling of the class name.  The misspelling is retained
+        as an alias.
 """
 
 from mazes.maze import Maze
 from mazes.Graphics.matplot_driver import Spider
 
-class Phocidae(Spider):
+class Pholcidae(Spider):
     """The daddy long legs of drivers
 
     The daddy long-legs spider has long somewhat clumsy legs.  This driver
@@ -46,11 +49,6 @@ class Phocidae(Spider):
     and west neighbors.  The passages are wide with no insets, perfect for
     the daddy long-legs to seek its prey.  Simplicity is the order of the
     day.
-
-        *** This is a misspelling ***
-
-    Apparently Phocidae is actually a family of seals.  Pholcidae is the
-    family of spiders which includes the daddy long-legs.
     """
 
     __slots__ = ("__maze", "__grid", "__fill")
@@ -141,5 +139,30 @@ class Phocidae(Spider):
         xs = [x, x+1, x+1, x]
         ys = [y, y, y+1, y+1]
         self.fill_polygon(xs, ys, color=color)
+
+class Phocidae(Pholcidae):
+    """The daddy long legs of drivers
+
+    The daddy long-legs spider has long somewhat clumsy legs.  This driver
+    works well for mazes defined on oblong grids provided all passages are
+    undirected and the only exits from a cell are to its north, south, east
+    and west neighbors.  The passages are wide with no insets, perfect for
+    the daddy long-legs to seek its prey.  Simplicity is the order of the
+    day.
+
+        *** This is a misspelling ***
+
+    Apparently Phocidae is actually a family of seals.  Pholcidae is the
+    family of spiders which includes the daddy long-legs.
+
+    This class is maintained for backward compatibility.
+    """
+
+    def __init__(self, maze:Maze):
+        """constructor"""
+        print("WARNING: The class name Phocidae is a misspelling.")
+        print("INFO: import Phocidae from 'mazes.Graphics.oblong1'.")
+        print("INFO: Please change the class name and import to Pholcidae.")
+        super().__init__(maze)
 
 # end module mazes.Graphics.oblong1